+++ application/Components/Auth.php
@@ -57,7 +57,7 @@ class Auth
 
         if ($user_id !== false && is_int($user_id) && $user_id > 0) {
             $user_id = (int)$user_id;
-            $curuser = container()->get('site')->getUser($user_id);
+            $curuser = container()->get(\App\Entity\User\UserFactory::class)->getUserById($user_id);
             if ($curuser->getStatus() !== UserStatus::DISABLED) {  // user status shouldn't be disabled
                 return $curuser;
             }
@@ -92,7 +92,7 @@ class Auth
 
         // Verity $payload['jti'] is force expired or not, And check if it same with $payload['aud']
         // And if not match, it means user logout this session or change password....
-        $uid = container()->get('site')->getUserFactory()->getUserIdBySession($payload['jti']);
+        $uid = container()->get(Entity\User\UserFactory::class)->getUserIdBySession($payload['jti']);
         if ($uid != $payload['aud']) {
             return false;
         }
@@ -120,7 +120,7 @@ class Auth
             return false;
         }
 
-        $user_id = container()->get('site')->getUserFactory()->getUserIdByPasskey($passkey);
+        $user_id = container()->get(Entity\User\UserFactory::class)->getUserIdByPasskey($passkey);
         return $user_id > 0 ? $user_id : false;
     }