+++ framework/Http/Route/Exception/NotFoundException.php
@@ -10,8 +10,6 @@ declare(strict_types=1);
 
 namespace Rid\Http\Route\Exception;
 
-
 class NotFoundException extends \RuntimeException
 {
-
 }